Rename the timed stream file
