+++ src/timed-stream.ts
